third_party/blink/web_tests/external/wpt/css/css-fonts/variations/at-font-face-descriptors-expected.txt

This is a testharness.js-based test.
Found 5 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] font-weight(valid): Simple range with equal upper and lower bounds: 500 500
  assert_equals: Unexpected resulting value. expected "500" but got "500 500"
[FAIL] font-stretch(valid): Simple range with equal upper and lower bounds: 100% 100%
  assert_equals: Unexpected resulting value. expected "100%" but got "100% 100%"
[FAIL] font-style(valid): Simple range with equal upper and lower bounds: oblique 10deg 10deg
  assert_equals: Unexpected resulting value. expected "oblique 10deg" but got "oblique 10deg 10deg"
Harness: the test ran to completion.
